Avoid memory allocation if `SmallSet` is small
A `SmallSet` is now allocation free when it has 2 or fewer elements.
This covers the majority of its uses since most nodes touch one or no
groups/ counters.

Before:
```
Benchmark 1: ./mine-highlight.out /home/tom/Programming/jit/benchmark/resources/llvm/llvm-100.txt
  Time (mean ± σ):       1.6 ms ±   0.2 ms    [User: 1.0 ms, System: 0.6 ms]
  Range (min … max):     1.5 ms …   2.5 ms    1352 runs
```

After:
```
Benchmark 1: ./mine-highlight.out /home/tom/Programming/jit/benchmark/resources/llvm/llvm-100.txt
  Time (mean ± σ):       1.4 ms ±   0.1 ms    [User: 0.8 ms, System: 0.5 ms]
  Range (min … max):     1.2 ms …   2.1 ms    2023 runs
```

+++ src/private/nfa.hpp
@@ -49,9 +49,9 @@ struct Condition {
 struct Node;
 struct Edge {
   size_t output_index;
-  SmallSet<size_t> start_groups;
-  SmallSet<size_t> end_groups;
-  SmallSet<size_t> counters;
+  SmallSet<uint32_t> start_groups;
+  SmallSet<uint32_t> end_groups;
+  SmallSet<uint32_t> counters;
 };
 
 struct Node {
